app/src/test/java/leetcode/ValidSudoku/SolutionTest.java
package leetcode.ValidSudoku;

import org.junit.Before;
import org.junit.Test;

import static org.junit.Assert.*;

public class SolutionTest {
    Solution isValidSudokuTest;
    char[][] board;
    boolean result;

    @Before
    public void init() {
        isValidSudokuTest = new Solution();
    }


    @Test
    public void Example1() {
        board = new char[][] {{'5','3','.','.','7','.','.','.','.'},{'6','.','.','1','9','5','.','.','.'},{'.','9','8','.','.','.','.','6','.'},{'8','.','.','.','6','.','.','.','3'},{'4','.','.','8','.','3','.','.','1'},{'7','.','.','.','2','.','.','.','6'},{'.','6','.','.','.','.','2','8','.'},{'.','.','.','4','1','9','.','.','5'},{'.','.','.','.','8','.','.','7','9'}};
        result = isValidSudokuTest.isValidSudoku(board);

        assertTrue(result);
    }

    @Test
    public void Example2() {
        board = new char[][] {{'8','3','.','.','7','.','.','.','.'},{'6','.','.','1','9','5','.','.','.'},{'.','9','8','.','.','.','.','6','.'},{'8','.','.','.','6','.','.','.','3'},{'4','.','.','8','.','3','.','.','1'},{'7','.','.','.','2','.','.','.','6'},{'.','6','.','.','.','.','2','8','.'},{'.','.','.','4','1','9','.','.','5'},{'.','.','.','.','8','.','.','7','9'}};
        result = isValidSudokuTest.isValidSudoku(board);

        assertFalse(result);
    }

}
